chore: refactor SSR CSR fetchs, and add endpoints
