+++ auth/services/server/registerUser.ts
